OAuth

+++ server/src/routes/memories.ts
@@ -3,16 +3,16 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  // app.addHook("preHandler", async (request) => {
-  //   await request.jwtVerify();
-  // });
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
 
   // GET ALL MEMORIES
   app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
-      // where: {
-      //   userId: request.user.sub,
-      // },
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -42,9 +42,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     });
 
-    // if (!memory.isPublic && memory.userId !== request.user.sub) {
-    //   return reply.status(401).send();
-    // }
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     return memory;
   });
@@ -64,7 +64,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "f65c4317-9c09-498d-9cf1-7e59300b21d3",
+        userId: request.user.sub,
       },
     });
 
@@ -72,7 +72,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
   });
 
   // UPDATE A MEMORY
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -87,11 +87,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-    // if (memory.userId !== request.user.sub) {
-    //   return reply.status(401).send();
-    // }
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
 
-    const memory = await prisma.memory.update({
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -106,21 +112,27 @@ export async function memoriesRoutes(app: FastifyInstance) {
   });
 
   // DELETE A MEMORY
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(request.params);
 
-    await prisma.memory.delete({
+    const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
     });
 
-    // if (memory.userId !== request.user.sub) {
-    //   return reply.status(401).send();
-    // }
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    await prisma.memory.delete({
+      where: {
+        id,
+      },
+    });
   });
 }
